Adjust code after test with single pecision
Exponent precomputation on dr variable yields a different result on GPU than on CPU if single precision is used. This is due to the implementation of exp function on GPUs and cannot be fixed. The error accumulates for longer-maturity bonds, but differences can be observed on any option.

Re-include getopt library in Common project