Fix reordering in gpu-outer and gpu-flat kernels
Reordering is not done anymore on all input data vectors, just on iteration space using one vector, ValuationIndices, to store the sorted indices. Only Widths and Heights vectors are reordered, the original index needs to be used when indexing them.

Fix and unify how lastUsedCIdx is used to get a new lastCStep in Seq and gpu-outer. Fix former out-of-bounds accesses to CashflowsSteps vector.